tests: move mock requests to conftest.py
pytest loads conftest.py automatically, so it's a good way to share
fixtures across tests.
